requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Data processing
openpyxl>=3.1.0
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0

# Utilities
click>=8.1.0
rich>=13.5.0
tqdm>=4.66.0

# Web scraping
requests>=2.31.0
certifi>=2023.0.0  # SSL certificates for better SSL error handling
beautifulsoup4>=4.12.0
lxml>=4.9.0
scrapy>=2.13.0  # For Modular11 scraper
twisted>=25.0.0  # Required by Scrapy

# Machine Learning (for rankings ML layer)
xgboost>=2.0.0
scikit-learn>=1.3.0

# Dashboard
streamlit>=1.28.0
matplotlib>=3.7.0
